Reuse the same VO conversion logic for the Web interface and RESTful API
Move GeneralSearchController's VO conversion logic into SearchService so
that it can be reused for the RESTful API.

Add a new ValueObjectConverter that implements Spring Converter
interface to apply VO conversion logic defined in BaseVoEnabledService.
This is used to rewrite the logic of converting VOs.

Load entities in the Compass search source so that they can be readily
available for conversion to value object without having to reload data
systematically. This addresses #278 as missing related entities is a
consequence of how entities are stored in Compass' index.

Simplify SearchService by moving fillObjects and webSpeedSearch in
SearchSettings.